The TinyDTLS defualt example upgraded.
Everything is in english now.
The confusing, and probably, legacy configuration from the
Makefile were removed.

The Project-Conf.h  is similar to the CoAPS.
A brief explanation of the situation for testing  the server
from extern clients is added.

P.d. In the branch CoAPS the server is unable to completed the
connection to an extern client (So, yeah, Contiki's DTLS is
faulty).