Upadting debug.c
For this branch (Which in theory is using the most updated code
from TinyDTLS).

The client seem to be hanged in TinyDTLS processes as unable to
be added to the 6BR's neighbor table.

Extern clients (using Eclipsion code) are unable to continue the
3rd flight as is stopped in the Second Hello crash the server
mote.

P.d. And exactly today, Sourceforge is not anymore accesible and
the Eclipse's code has been updated (Future work is to patch
the motes code of TinyDTLS).